Removing UiImage from NodeBundle, touching up docs, and separating camera config stuff from the pre-made UI node bundles.
